Refactoring ChangesToDoS class, moving func(s)
