+++ application/service_sync/changes_to_dos.py
@@ -50,187 +50,3 @@ class ChangesToDoS:
     # new_day_opening_times e.g. new_monday_opening_times
     # current_day_opening_times e.g. current_monday_opening_times
     # The type of the value is a list of OpenPeriod objects
-
-    def check_for_standard_opening_times_day_changes(self, weekday: str) -> bool:
-        """Check if the standard opening times have changed for a specific day.
-
-        Args:
-            weekday (str): The day of the week lowercase to check  (e.g. "monday")
-
-        Returns:
-            bool: If there are changes to the standard opening times (not valiated)
-        """
-        dos_standard_open_dates: StandardOpeningTimes = self.dos_service.standard_opening_times
-        nhs_standard_open_dates: StandardOpeningTimes = self.nhs_entity.standard_opening_times
-        dos_opening_times = dos_standard_open_dates.get_openings(weekday)
-        nhs_opening_times = nhs_standard_open_dates.get_openings(weekday.title())
-        if not dos_standard_open_dates.same_openings(nhs_standard_open_dates, weekday):
-            logger.info(
-                f"{weekday.title()} opening times not equal. "
-                f"dos={opening_period_times_from_list(dos_opening_times)}, "
-                f"nhs={opening_period_times_from_list(nhs_opening_times)}",
-            )
-            # Set variable for the correct day
-            setattr(self, f"current_{weekday}_opening_times", dos_opening_times)
-            setattr(self, f"new_{weekday}_opening_times", nhs_opening_times)
-            return True
-        else:
-            logger.info(
-                f"{weekday.title()} opening times are equal, so no change. "
-                f"dos={opening_period_times_from_list(dos_opening_times)} "
-                f"nhs={opening_period_times_from_list(nhs_opening_times)}",
-            )
-            return False
-
-    def check_for_specified_opening_times_changes(self) -> bool:
-        """Check if the specified opening times have changed
-        Also past specified opening times are removed from the comparison.
-
-        Returns:
-            bool: If there are changes to the specified opening times (not valiated)
-        """
-        dos_spec_open_dates = self.dos_service.specified_opening_times
-        nhs_spec_open_dates = self.nhs_entity.specified_opening_times
-        future_nhs_spec_open_dates = SpecifiedOpeningTime.remove_past_dates(self.nhs_entity.specified_opening_times)
-        if len(nhs_spec_open_dates) != len(future_nhs_spec_open_dates):
-            logger.info(
-                "Removing Specified opening times that occur in the past",
-                extra={"all_nhs": nhs_spec_open_dates, "future_nhs": future_nhs_spec_open_dates},
-            )
-        if SpecifiedOpeningTime.equal_lists(dos_spec_open_dates, future_nhs_spec_open_dates):
-            logger.info(
-                "Specified opening times are equal, so no change",
-                extra={"dos": dos_spec_open_dates, "nhs": future_nhs_spec_open_dates},
-            )
-            return False
-        else:
-            logger.info(
-                "Specified opening times not equal",
-                extra={"dos": dos_spec_open_dates, "nhs": future_nhs_spec_open_dates},
-            )
-            self.current_specified_opening_times = dos_spec_open_dates
-            self.new_specified_opening_times = future_nhs_spec_open_dates
-            return True
-
-    def check_for_address_and_postcode_for_changes(self) -> tuple[bool, bool, DoSLocation | None]:
-        """Check if address and postcode have changed between dos_service and nhs_entity,
-        Postcode changes are validated against the DoS locations table.
-
-        Returns:
-            Tuple[bool, bool]: Tuple of booleans, first is if address has changed, second is if postcode has changed, third is the DoSLocation object for the postcode
-        """  # noqa: E501
-        before_title_case_address = self.nhs_entity.address_lines
-        self.nhs_entity.address_lines = list(map(format_address, self.nhs_entity.address_lines))
-        logger.info(
-            f"Address after title casing: {self.nhs_entity.address_lines}",
-            extra={"before": before_title_case_address, "after": self.nhs_entity.address_lines},
-        )
-        nhs_uk_address_string = "$".join(self.nhs_entity.address_lines)
-        dos_address = self.dos_service.address
-        is_address_same = True
-        if dos_address != nhs_uk_address_string:
-            is_address_same = False
-            logger.info(f"Address is not equal, {dos_address=} != {nhs_uk_address_string=}")
-            self.new_address = nhs_uk_address_string
-            self.current_address = dos_address
-        else:
-            logger.info(f"Address is equal, {dos_address=} == {nhs_uk_address_string=}")
-
-        dos_postcode = self.dos_service.normal_postcode()
-        nhs_postcode = self.nhs_entity.normal_postcode()
-        is_postcode_same = True
-        valid_dos_location = None
-        if dos_postcode != nhs_postcode:
-            logger.info(f"Postcode is not equal, {dos_postcode=} != {nhs_postcode=}")
-            valid_dos_location = get_valid_dos_location(nhs_postcode)
-            valid_dos_postcode = valid_dos_location.postcode if valid_dos_location else None
-            if valid_dos_postcode is None:
-                log_invalid_nhsuk_postcode(self.nhs_entity, self.dos_service)  # type: ignore
-                if not is_address_same:
-                    is_address_same = True
-                    self.new_address = None
-                    self.current_address = None
-                    logger.info("Deleted address change as postcode is invalid")
-            else:
-                if is_address_same:
-                    logger.info(f"Address is equal but Postcode is not equal, {dos_postcode=} != {nhs_postcode=}")
-                self.new_postcode = valid_dos_postcode
-                self.current_postcode = self.dos_service.postcode
-                is_postcode_same = False
-        else:
-            logger.info(f"Postcode are equal, {dos_postcode=} == {nhs_postcode=}")
-        return not is_address_same, not is_postcode_same, valid_dos_location
-
-    def check_website_for_change(self) -> bool:
-        """Compares the website of from the dos_service and nhs_entity."""
-        if is_val_none_or_empty(self.nhs_entity.website) and not is_val_none_or_empty(self.dos_service.web):
-            # Deleting the existing website
-            self.current_website = self.dos_service.web
-            self.new_website = None
-            return True
-        elif self.nhs_entity.website is not None and self.nhs_entity.website != "":
-            self.current_website = self.dos_service.web
-            # Adding a new website
-            nhs_uk_website = format_website(self.nhs_entity.website)
-            self.new_website = nhs_uk_website
-            return self.compare_and_validate_website(self.dos_service, self.nhs_entity, nhs_uk_website)
-        return False
-
-    def compare_and_validate_website(self, dos_service: DoSService, nhs_entity: NHSEntity, nhs_website: str) -> bool:
-        """Compares the website of from the dos_service and formatted nhs website.
-
-        Args:
-            dos_service (DoSService): DoSService object to compare
-            nhs_entity (NHSEntity): NHS Entity object to compare
-            nhs_website (str): Formatted NHS website
-
-        Returns:
-            bool: True if the website has changed and is valid, False if not
-        """
-        dos_website = dos_service.web
-        if dos_website != nhs_website:
-            logger.info(f"Website is not equal, {dos_website=} != {nhs_website=}")
-            # Regular expression to match DoS's websites check
-            return validate_website(nhs_entity, nhs_website)
-        else:
-            logger.info(f"Website is equal, {dos_website=} == {nhs_website=}")
-        return False
-
-    def check_public_phone_for_change(self) -> bool:
-        """Compares the public phone of from the dos_service and nhs_entity.
-
-        Returns:
-            bool: True if the public phone has changed, False if not
-        """
-        self.current_public_phone = self.dos_service.publicphone
-        self.new_public_phone = self.nhs_entity.phone
-        if str(self.current_public_phone) != str(self.new_public_phone) and (
-            not is_val_none_or_empty(self.current_public_phone) or not is_val_none_or_empty(self.new_public_phone)
-        ):
-            logger.info(
-                f"Public Phone is not equal, DoS='{self.current_public_phone}' != NHS UK='{self.new_public_phone}'",
-            )
-            return True
-        else:
-            logger.info(f"Public Phone is equal, DoS='{self.current_public_phone}' == NHS UK='{self.new_public_phone}'")
-            return False
-
-    def check_palliative_care_for_change(self) -> bool:
-        """Compares the palliative care of from the dos_service and nhs_entity.
-
-        Returns:
-            bool: True if the palliative care is different, False if not
-        """
-        self.current_palliative_care = self.dos_service.palliative_care
-        self.new_palliative_care = self.nhs_entity.palliative_care
-        if self.current_palliative_care != self.new_palliative_care:
-            logger.info(
-                f"Palliative Care is not equal, DoS='{self.current_palliative_care}' "
-                + f"!= NHS UK='{self.new_palliative_care}'",
-            )
-            return True
-        else:
-            logger.info(
-                f"Palliative Care is equal, DoS='{self.current_palliative_care}' == NHSUK='{self.new_palliative_care}'",
-            )
-            return False
